Add files via upload
I wasn't sure how to save these VBA script files. They should contain the same information. My Mac wants to upload it as a Basic file with the extension .bas. I have also tried saving it as a .vb file. It contains a sub that loops through all worksheets. That sub calls the StockAnalysis sub, which calls the GreatestOnStockAnalsis sub.